test-requirements.txt
coverage
django-recaptcha
pexpect==4.8.0
pyvirtualdisplay
selenium
